% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_func.R
\name{ISTA.main}
\alias{ISTA.main}
\title{Main function}
\usage{
ISTA.main(data.X, data.Y, scale = T, lambda = 0.01, method = NULL,
  epoch = 10000, patience = 50)
}
\arguments{
\item{data.X}{Predictor matrix}

\item{data.Y}{Response matrix}

\item{scale}{Whether the matrix should be scaled when applying algrithm, default is TRUE}

\item{lambda}{Parameter of LASSO regression, can be a vector with length > 1, default is 1e-2}

\item{method}{Method used to select optimal parameters, can only chosen from 'CV', 'AIC' or 'BIC', default is NULL}

\item{epoch}{The maximum number of steps allowed when training the model, default is 1e4}

\item{patience}{Used to decide when end training after a minima occured, default is 50}
}
\value{
MSE between prediction values and true values of response'
}
\description{
Iterative Soft_thresholding Algorithm (ISTA)
}
\details{
ISTA algrithm combined with Backtracking line search, using Cross-validation/AIC/BIC to select optimal parameter
}
\author{
Yuan Zhou, Lingsong Meng
}
